backend/requirements-migrations.txt
SQLAlchemy>=1.4
psycopg2-binary>=2.9
alembic>=1.11
python-dotenv>=1.0
pandas>=2.0